Rename : 변경,  AutoAppConfig 프로젝트 시작 루트에 두기 위해 basePackages를 생략하기 위해 파일 이동
- discount, order 파일 이동
- AutoAppConfig : com.example.core -> 프로젝트 시작 루트
- com.exmple.core.discount
- com.exmple.core.member
- com.exmple.core.order

프로젝트 시작루트인 com.example.core에 설정 정보를 두고 basePackages는 생략한다.
