+++ src/main/java/com/example/core/AutoAppConfig.java
@@ -7,6 +7,8 @@ import org.springframework.context.annotation.FilterType;
 @Configuration
 
 @ComponentScan(
+        basePackages = "com.example.core.member",
+        basePackageClasses = AutoAppConfig.class,
         excludeFilters = @ComponentScan.Filter(type = FilterType.ANNOTATION,classes = Configuration.class)
 )
 public class AutoAppConfig {
