+++ src/main/java/com/example/core/discoount/RateDiscountPolicy.java
@@ -1,16 +1,19 @@
-package discount;
+package com.example.core.discoount;
 
 import com.example.core.member.Grade;
 import com.example.core.member.Member;
 
-public class FixDiscountPolicy  implements DisCountPolicy{
+import org.springframework.stereotype.Component;
 
-    private  int discountFixAmount = 1000;
+@Component
+public class RateDiscountPolicy implements DisCountPolicy{
+
+   private  int discountPercent = 10;
 
     @Override
     public int discount(Member member, int price) {
         if(member.getGrade() == Grade.VIP){
-            return discountFixAmount;
+            return price * discountPercent/100;
         }else{
             return 0;
         }
