+++ src/main/java/com/example/core/order/Order.java
@@ -1,4 +1,4 @@
-package order;
+package com.example.core.order;
 
 public class Order {
