+++ src/test/java/discount/RateDiscountPolicyTest.java
@@ -1,13 +1,12 @@
 package discount;
 
+import com.example.core.discoount.RateDiscountPolicy;
 import com.example.core.member.Grade;
 import com.example.core.member.Member;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 class RateDiscountPolicyTest {
